#117 - reducing class amount (#120)
