--- trade-module/src/main/java/io/edpn/backend/trade/application/service/ReceiveStationArrivalDistanceResponseService.java
@@ -1,46 +1,0 @@
-package io.edpn.backend.trade.application.service;
-
-import io.edpn.backend.messageprocessorlib.application.dto.eddn.data.StationArrivalDistanceResponse;
-import io.edpn.backend.trade.application.domain.Station;
-import io.edpn.backend.trade.application.domain.System;
-import io.edpn.backend.trade.application.port.incomming.kafka.ReceiveKafkaMessageUseCase;
-import io.edpn.backend.trade.application.port.outgoing.station.LoadOrCreateBySystemAndStationNamePort;
-import io.edpn.backend.trade.application.port.outgoing.station.UpdateStationPort;
-import io.edpn.backend.trade.application.port.outgoing.stationarrivaldistancerequest.DeleteStationArrivalDistanceRequestPort;
-import io.edpn.backend.trade.application.port.outgoing.system.LoadOrCreateSystemByNamePort;
-import lombok.RequiredArgsConstructor;
-import lombok.extern.slf4j.Slf4j;
-
-import java.util.concurrent.CompletableFuture;
-
-@RequiredArgsConstructor
-@Slf4j
-public class ReceiveStationArrivalDistanceResponseService implements ReceiveKafkaMessageUseCase<StationArrivalDistanceResponse> {
-
-    private final LoadOrCreateSystemByNamePort loadOrCreateSystemByNamePort;
-    private final LoadOrCreateBySystemAndStationNamePort loadOrCreateBySystemAndStationNamePort;
-    private final DeleteStationArrivalDistanceRequestPort deleteStationArrivalDistanceRequestPort;
-    private final UpdateStationPort updateStationPort;
-
-    @Override
-    public void receive(StationArrivalDistanceResponse message) {
-        String systemName = message.systemName();
-        String stationName = message.stationName();
-        double arrivalDistance = message.arrivalDistance();
-
-        CompletableFuture<System> systemCompletableFuture = CompletableFuture.supplyAsync(() -> loadOrCreateSystemByNamePort.loadOrCreateSystemByName(systemName));
-
-        // get station
-        CompletableFuture<Station> stationCompletableFuture = CompletableFuture.supplyAsync(() -> loadOrCreateBySystemAndStationNamePort.loadOrCreateBySystemAndStationName(systemCompletableFuture.join(), stationName));
-        stationCompletableFuture.whenComplete((station, throwable) -> {
-            if (throwable != null) {
-                log.error("Exception occurred in retrieving station", throwable);
-            } else {
-                station.setArrivalDistance(arrivalDistance);
-            }
-        });
-
-        updateStationPort.update(stationCompletableFuture.join());
-        deleteStationArrivalDistanceRequestPort.delete(systemName, stationName);
-    }
-}
