+++ trade-module/src/test/java/io/edpn/backend/trade/application/service/CleanUpObsoleteStationArrivalDistanceRequestsUseCaseTest.java
@@ -1,19 +1,30 @@
 package io.edpn.backend.trade.application.service;
 
+import com.fasterxml.jackson.databind.ObjectMapper;
 import io.edpn.backend.messageprocessorlib.application.dto.eddn.data.StationDataRequest;
 import io.edpn.backend.trade.application.domain.Station;
 import io.edpn.backend.trade.application.domain.System;
 import io.edpn.backend.trade.application.domain.filter.FindStationFilter;
+import io.edpn.backend.trade.application.dto.web.object.mapper.MessageMapper;
+import io.edpn.backend.trade.application.port.outgoing.kafka.SendKafkaMessagePort;
+import io.edpn.backend.trade.application.port.outgoing.station.LoadOrCreateBySystemAndStationNamePort;
 import io.edpn.backend.trade.application.port.outgoing.station.LoadStationsByFilterPort;
+import io.edpn.backend.trade.application.port.outgoing.station.UpdateStationPort;
+import io.edpn.backend.trade.application.port.outgoing.stationarrivaldistancerequest.CleanUpObsoleteStationArrivalDistanceRequestsUseCase;
+import io.edpn.backend.trade.application.port.outgoing.stationarrivaldistancerequest.CreateStationArrivalDistanceRequestPort;
 import io.edpn.backend.trade.application.port.outgoing.stationarrivaldistancerequest.DeleteStationArrivalDistanceRequestPort;
+import io.edpn.backend.trade.application.port.outgoing.stationarrivaldistancerequest.ExistsStationArrivalDistanceRequestPort;
 import io.edpn.backend.trade.application.port.outgoing.stationarrivaldistancerequest.LoadAllStationArrivalDistanceRequestsPort;
+import io.edpn.backend.trade.application.port.outgoing.system.LoadOrCreateSystemByNamePort;
 import java.util.Collections;
 import java.util.List;
+import java.util.concurrent.Executor;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
 import org.mockito.Mock;
 import org.mockito.junit.jupiter.MockitoExtension;
+import org.springframework.retry.support.RetryTemplate;
 
 import static org.hamcrest.MatcherAssert.assertThat;
 import static org.hamcrest.Matchers.equalTo;
@@ -28,18 +39,49 @@ public class CleanUpObsoleteStationArrivalDistanceRequestsUseCaseTest {
 
     @Mock
     private LoadStationsByFilterPort loadStationsByFilterPort;
-
     @Mock
     private LoadAllStationArrivalDistanceRequestsPort loadAllStationArrivalDistanceRequestsPort;
-
     @Mock
     private DeleteStationArrivalDistanceRequestPort deleteStationArrivalDistanceRequestPort;
+    @Mock
+    private LoadOrCreateSystemByNamePort loadOrCreateSystemByNamePort;
+    @Mock
+    private LoadOrCreateBySystemAndStationNamePort loadOrCreateBySystemAndStationNamePort;
+    @Mock
+    private ExistsStationArrivalDistanceRequestPort existsStationArrivalDistanceRequestPort;
+    @Mock
+    private CreateStationArrivalDistanceRequestPort createStationArrivalDistanceRequestPort;
+    @Mock
+    private UpdateStationPort updateStationPort;
+    @Mock
+    private SendKafkaMessagePort sendKafkaMessagePort;
+    @Mock
+    private RetryTemplate retryTemplate;
+    @Mock
+    private Executor executor;
+    @Mock
+    private ObjectMapper objectMapper;
+    @Mock
+    private MessageMapper messageMapper;
 
-    private CleanUpObsoleteStationArrivalDistanceRequestsService underTest;
+    private CleanUpObsoleteStationArrivalDistanceRequestsUseCase underTest;
 
     @BeforeEach
     public void setUp() {
-        underTest = new CleanUpObsoleteStationArrivalDistanceRequestsService(loadStationsByFilterPort, loadAllStationArrivalDistanceRequestsPort, deleteStationArrivalDistanceRequestPort);
+        underTest = new StationArrivalDistanceInterModuleCommunicationService(
+                loadStationsByFilterPort,
+                loadAllStationArrivalDistanceRequestsPort,
+                loadOrCreateSystemByNamePort,
+                loadOrCreateBySystemAndStationNamePort,
+                existsStationArrivalDistanceRequestPort,
+                createStationArrivalDistanceRequestPort,
+                deleteStationArrivalDistanceRequestPort,
+                updateStationPort,
+                sendKafkaMessagePort,
+                retryTemplate,
+                executor,
+                objectMapper,
+                messageMapper);
     }
 
     @Test
@@ -48,7 +90,7 @@ public class CleanUpObsoleteStationArrivalDistanceRequestsUseCaseTest {
                 .hasArrivalDistance(false)
                 .build();
 
-        assertThat(RequestMissingStationArrivalDistanceService.FIND_STATION_FILTER, equalTo(findStationFilter));
+        assertThat(StationArrivalDistanceInterModuleCommunicationService.FIND_STATION_FILTER, equalTo(findStationFilter));
     }
 
     @Test
